+++ platform-core/src/main/java/com/webank/wecube/platform/core/parser/datamodel/DataModelExpressionListener.java
@@ -1,8 +1,8 @@
 package com.webank.wecube.platform.core.parser.datamodel;
 
-import com.webank.wecube.platform.core.dto.DataModelExpressionDto;
 import com.webank.wecube.platform.core.parser.datamodel.generated.DataModelBaseListener;
 import com.webank.wecube.platform.core.parser.datamodel.generated.DataModelParser;
+import com.webank.wecube.platform.core.support.parser.datamodel.DataModelExpressionDto;
 
 import java.util.LinkedList;
 import java.util.Queue;
